修改modbus 数据读取方式，modbus 一次性最多读取126个数据。分开读取数据。设置最大最小压力值，出入int，不要传入float值，设 置卸载压力和加载压差，同样传入int值，不传入float值。剩下历史数据的图表展示有点问题，需要修改。

+++ compressorsetdialog.h
@@ -18,7 +18,7 @@ public:
     ~CompressorSetDialog();
 
 signals:
-    void closeCurrentDialog(int compressorNo,float uninstallPressure,float pressureDiff);
+    void closeCurrentDialog(int compressorNo,int uninstallPressure,int pressureDiff);
 
 private slots:
     void on_setCompressorBtn_clicked();
